--- a/docker-compose.nodejs.yml
+++ b/docker-compose.nodejs.yml
@@ -26,3 +26,9 @@ services:
       - /var/www/dist
     ports:
       - '8080:8080'
+    networks:
+      - some-net
+
+networks:
+  some-net:
+    driver: bridge