--- a/docker/varnish/docker-compose copy/docker-compose.varnish.yml
+++ b/docker/varnish/docker-compose copy/docker-compose.varnish.yml
@@ -8,3 +8,9 @@ services:
       - ./docker/varnish/config.vcl:/usr/local/etc/varnish/default.vcl
     ports:
       - '1234:80'
+    networks:
+      - vuestorefrontapi_some-net
+
+networks:
+  vuestorefrontapi_some-net:
+    external: true