--- a/docker/varnish/docker-compose/docker-compose.nodejs.yml
+++ b/docker/varnish/docker-compose/docker-compose.nodejs.yml
@@ -26,4 +26,9 @@ services:
       - /var/www/dist
     ports:
       - '8080:8080'
-      
+    networks:
+      - some-net
+
+networks:
+  some-net:
+    driver: bridge