--- a/src/api/catalog.ts
+++ b/src/api/catalog.ts
@@ -129,7 +129,15 @@ export default ({config, db}) => async function (req, res, body) {
         if (entityType === 'product') {
           resultProcessor.process(_resBody.hits.hits, groupId).then(async (result) => {
             _resBody.hits.hits = result
-            _cacheStorageHandler(config, _resBody, reqHash, tagsArray)
+            if (config.get('varnish.enabled')) {
+              // Add tags to cache, so we can display them in response headers then
+              _cacheStorageHandler(config, {
+                ..._resBody,
+                tags: tagsArray
+              }, reqHash, tagsArray)
+            } else {
+              _cacheStorageHandler(config, _resBody, reqHash, tagsArray)
+            }
             if (_resBody.aggregations && config.entities.attribute.loadByAttributeMetadata) {
               const attributeListParam = AttributeService.transformAggsToAttributeListParam(_resBody.aggregations)
               // find attribute list
@@ -143,7 +151,15 @@ export default ({config, db}) => async function (req, res, body) {
         } else {
           resultProcessor.process(_resBody.hits.hits).then((result) => {
             _resBody.hits.hits = result
-            _cacheStorageHandler(config, _resBody, reqHash, tagsArray)
+            if (config.get('varnish.enabled')) {
+              // Add tags to cache, so we can display them in response headers then
+              _cacheStorageHandler(config, {
+                ..._resBody,
+                tags: tagsArray
+              }, reqHash, tagsArray)
+            } else {
+              _cacheStorageHandler(config, _resBody, reqHash, tagsArray)
+            }
             res.json(_outputFormatter(_resBody, responseFormat));
           }).catch((err) => {
             console.error(err)
@@ -161,6 +177,11 @@ export default ({config, db}) => async function (req, res, body) {
     ).then(output => {
       if (output !== null) {
         res.setHeader('X-VS-Cache', 'Hit')
+        if (config.get('varnish.enabled')) {
+          const tagsHeader = output.tags.join(' ')
+          res.setHeader('X-VS-Cache-Tags', tagsHeader)
+          delete output.tags
+        }
         res.json(output)
         console.log(`cache hit [${req.url}], cached request: ${Date.now() - s}ms`)
       } else {