@Fifciu 2020-02-15
I've rebased it to the new develop. Removed `network` stuff from docker composes and put configured ones in docker/varnish/docker-compose. Varnish will only work if config.varnish.enabled equals true